requirements.txt
jupyter==1.0.0
matplotlib==3.1.1
numpy==1.21.6
torch==1.13.1
scipy==1.7.3
tqdm==4.36.1
pillow==9.5.0
scikit-image==0.19.3
ipython_genutils==0.2.0
pyyaml==6.0.1
